requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0Django==3.0.5
django-crispy-forms==1.9.0
django-summernote==0.8.11.6
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
